scripts/faux_event/requirements.txt
blinker==1.9.0
cachecontrol==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudevents==1.11.0
cryptography==44.0.2
deprecation==2.1.0
firebase-admin==6.7.0
flask==3.1.0
functions-framework==3.8.2
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-cloud-tasks==2.19.2
google-crc32c==1.7.1
google-events==0.14.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
msgpack==1.1.0
nodeenv==1.9.1
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pyfarmhash==0.4.0
pyjwt==2.10.1
pyparsing==3.2.3
pyright==1.1.399
requests==2.32.3
rsa==4.9.1
typing-extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
watchdog==6.0.0
werkzeug==3.1.3
